Switch from delayed to spawn, reorg
